Code restructure & heartbeat timeout improvement
